SnakeGame/src/test/java/SnakeGameTest/SnakeTest.java

package SnakeGameTest;

import com.mycompany.snakegame.Snake;
import java.util.ArrayList;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

public class SnakeTest {
    
    private Snake snake;
    
    @Before
    public void setUp() {
        this.snake = new Snake(1,1);
    }
    
    @Test
    public void isConstructorhRight() {
        ArrayList<Integer> list = new ArrayList<>();
        list.add(1);
        assertEquals(list, snake.getX());
        assertEquals(list, snake.getY());
        assertEquals(1,snake.getLength());
        assertEquals(0,snake.getHead());
    }
    
    @Test
    public void isSnakeGrowing() {
    
    }

}
